refactor: split data utilities by domain
