+++ src/components/RecentPosts.astro
@@ -1,7 +1,7 @@
 ---
 import BlogCard from "@/components/BlogCard.astro";
 import Link from "@/components/Link.astro";
-import { getAllBlogPosts } from "@/lib/data-utils";
+import { getAllBlogPosts } from "@/lib/blog";
 import { cn } from "@/lib/utils";
 import { ArrowRight } from "lucide-astro";
